fix: npm login
previously the login on publish was broken with new versions of npm.
Now it will launch a terminal window for the user to login, which will launch the login flow.
Also it now distributes and installs npm as necessary so we no longer need npm in order to install it.
next we can provide it as a full gui desktop app.