feat: Add support for generateLegacyBundles (#229)
* feat: Add support for generateLegacyBundles

This is a property to be added to the jdeploy object of the package.json.
true (default) means that when generating windows and linux installer bundles, it will generate them with both new and old naming conventions.  The new naming convention is to include the architecture in the name.  The old naming convention is not to.
false means that we should not generate duplicate bundles.  Just use new naming conventions.

This defaults to true right now so we don't break existing automations, but will be switched to false in a later version.
When generating new projects, this is now explicitly set to false.
It will emit warnings to stdout when generating these duplicate bundles to alert users to them, and describe how to disable them.

* disable test on windows due to file locking

* disabled test on windows - needed JUnit 4 notation

* disable package service integration test on windows

it chokes on temp file deletion

* https://github.com/shannah/jdeploy/issues/227